♻️ Refactor: #12 코드 리팩토링 진행
코드 리팩토링을 진행하였습니다.
