+++ NLP/NLP/Sources/Dialog/DialogManager.swift
@@ -26,7 +26,7 @@ class DialogManager: ObservableObject {
                     // MARK: computer instruction 변경(암호를 맞춘 후, 컴퓨터가 어떠한 응답을 내뱉어줄지에 대한 instruction으로) 및 세션 초기화
                     self?.initializeSession(
                         dialogPartner: .computer,
-                        instructions: ConstantInstruction.computerOnboarding,
+                        instructions: ConstantInstructions.computerOnboarding,
                         tools: [] // TODO: 미결정
                     )
                 })
