+++ NLP/NLP/Sources/Utils/Constants/ConstantNodeType.swift
@@ -1,11 +1,11 @@
 //
-//  NodeType.swift
+//  ConstantNodes.swift
 //  NLP
 //
 //  Created by Ted on 7/14/25.
 //
 
-struct NodeType {
+struct ConstantNodes {
     /// Sprite Scene 내에서 미리 설정된 노드 이름 "TileNodes"를 나타냅니다.
     static var tileNodes = "TileNodes"
     /// Sprite Scene 내에서 미리 설정된 노드 이름 "Floor"를 나타냅니다.
