+++ NLP/NLP/Sources/Utils/Constants/ConstantScreenSize.swift
@@ -1,0 +1,13 @@
+//
+//  ConstantScreenSize.swift
+//  NLP
+//
+//  Created by Ted on 7/14/25.
+//
+
+import SwiftUI
+
+struct ConstantScreenSize {
+    static var screenWidth = UIScreen.main.bounds.width
+    static var screenHeight = UIScreen.main.bounds.height
+}
